fsl_qman: fix the wrong type in ceetm statistic query structure
Reviewed-on: http://git.am.freescale.net:8181/5121
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>

+++ include/linux/fsl_qman.h
@@ -873,7 +873,7 @@ struct qm_mcc_ceetm_statistics_query_write {
 	u8 ct;
 	u8 __reserved2[13];
 	u64 frm_cnt:40;
-	u16 __reserved3[2];
+	u8 __reserved3[2];
 	u64 byte_cnt:48;
 	u8 __reserved[32];
 } __packed;
